requirements.txt
Flask
PyMuPDF
python-docx
pandas
pillow
pytesseract
numpy
transformers